+++ resources/configuration.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: mariadb
-# Resource:: extraconf
+# Resource:: configuration
 #
 
 actions :add, :remove
@@ -9,4 +9,4 @@ default_action :add
 # name of the extra conf file, used for .cnf filename
 attribute :conf_name, kind_of: String, name_attribute: true
 attribute :section, kind_of: String
-attribute :option,  kind_of: Array, default: []
+attribute :option,  kind_of: Hash, default: {}